refactor: split service/ into service contracts + provider implementations
Separate the two axes that service/ was conflating: the service kinds
(llm/stt/tts contracts + shared base processors) stay under service/,
and the 20 vendor packages move to provider/. A provider implements one
or more services (openai: LLM+STT+TTS, deepgram: STT+TTS, groq: LLM+STT),
so grouping by provider keeps each vendor cohesive.

Package names are unchanged; only import paths move
(service/<vendor> -> provider/<vendor>). Wrapper and example imports
updated; PLAN.md layout refreshed. No behavior change.
