+++ examples/voicebot/main.go
@@ -29,20 +29,20 @@ import (
 	"github.com/gojargo/jargo/pipeline"
 	"github.com/gojargo/jargo/processor"
 	"github.com/gojargo/jargo/rtvi"
-	"github.com/gojargo/jargo/service/anthropic"
-	"github.com/gojargo/jargo/service/assemblyai"
-	"github.com/gojargo/jargo/service/cartesia"
-	"github.com/gojargo/jargo/service/deepgram"
-	"github.com/gojargo/jargo/service/deepseek"
-	"github.com/gojargo/jargo/service/elevenlabs"
-	"github.com/gojargo/jargo/service/gladia"
-	"github.com/gojargo/jargo/service/google"
-	"github.com/gojargo/jargo/service/groq"
-	"github.com/gojargo/jargo/service/lmnt"
-	"github.com/gojargo/jargo/service/ollama"
-	"github.com/gojargo/jargo/service/openai"
-	"github.com/gojargo/jargo/service/rime"
-	"github.com/gojargo/jargo/service/together"
+	"github.com/gojargo/jargo/provider/anthropic"
+	"github.com/gojargo/jargo/provider/assemblyai"
+	"github.com/gojargo/jargo/provider/cartesia"
+	"github.com/gojargo/jargo/provider/deepgram"
+	"github.com/gojargo/jargo/provider/deepseek"
+	"github.com/gojargo/jargo/provider/elevenlabs"
+	"github.com/gojargo/jargo/provider/gladia"
+	"github.com/gojargo/jargo/provider/google"
+	"github.com/gojargo/jargo/provider/groq"
+	"github.com/gojargo/jargo/provider/lmnt"
+	"github.com/gojargo/jargo/provider/ollama"
+	"github.com/gojargo/jargo/provider/openai"
+	"github.com/gojargo/jargo/provider/rime"
+	"github.com/gojargo/jargo/provider/together"
 	"github.com/gojargo/jargo/transport"
 	"github.com/gojargo/jargo/transport/pionrtc"
 	"github.com/gojargo/jargo/turntaking"
