+++ provider/cerebras/cerebras.go
@@ -1,7 +1,7 @@
 // Package cerebras provides Cerebras's OpenAI-compatible LLM service.
 package cerebras
 
-import "github.com/gojargo/jargo/service/openai"
+import "github.com/gojargo/jargo/provider/openai"
 
 const (
 	baseURL      = "https://api.cerebras.ai/v1"
