+++ provider/fireworks/fireworks.go
@@ -1,7 +1,7 @@
 // Package fireworks provides Fireworks AI's OpenAI-compatible LLM service.
 package fireworks
 
-import "github.com/gojargo/jargo/service/openai"
+import "github.com/gojargo/jargo/provider/openai"
 
 const (
 	baseURL      = "https://api.fireworks.ai/inference/v1"
