+++ provider/openai/llm_tools_test.go
