+++ provider/openai/stt.go
