+++ provider/rime/rime.go
