+++ provider/xai/xai.go
@@ -1,7 +1,7 @@
 // Package xai provides xAI's (Grok) OpenAI-compatible LLM service.
 package xai
 
-import "github.com/gojargo/jargo/service/openai"
+import "github.com/gojargo/jargo/provider/openai"
 
 const (
 	baseURL      = "https://api.x.ai/v1"
